LayoutTests/fast/scrolling/ios/click-events-during-momentum-scroll-in-main-frame.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true AsyncOverflowScrollingEnabled=true AsyncFrameScrollingEnabled=true CSSScrollAnchoringEnabled=false ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<style>
html, body {
    width: 100%;
    margin: 0px;
    padding: 0px;
}

#content {
    width: 100px;
    height: 5000px;
    box-sizing: border-box;
    padding: 20px;
    background: linear-gradient(to bottom, red 0%, green 50%, blue 100%);
}
</style>
<body>
<div id="content"></div>
</body>
<script>
jsTestIsAsync = true;
const minimumExpectedScrollTop = 1200;

addEventListener("load", async () => {
    description("Verifies that tapping the page during momentum scrolling does not dispatch click events to the page. To run the test manually, swipe up to scroll down; while the page is momentum scrolling, tap the screen to interrupt scrolling. The page should not observe any click events.");
    document.body.addEventListener("click", () => testFailed("Observed unexpected click event."), { once: true });

    if (!window.testRunner)
        return;

    await UIHelper.dragFromPointToPoint(160, 450, 160, 0, 0.1);
    await shouldBecomeEqual("document.scrollingElement.scrollTop > minimumExpectedScrollTop", "true");
    await UIHelper.sendEventStream(new UIHelper.EventStreamBuilder()
        .begin(160, document.scrollingElement.scrollTop + 350)
        .wait(0.1)
        .end()
        .takeResult());

    await UIHelper.ensureStablePresentationUpdate();
    finishJSTest();
});
</script>
</html>
